feat: make classroom and venue to be clickable
